Refactor: split up `apid::web`
