+++ crates/apid/src/web/resources.rs
@@ -1,13 +1,8 @@
 use axum::extract::{Json, Path, State};
 use axum::http::StatusCode;
 use axum::response::IntoResponse;
-use axum::routing::{delete, get, patch, post};
-use axum::Router;
-use serde::Serialize;
 use serde_json::Value;
 use std::collections::HashMap;
-use std::net::Ipv4Addr;
-use std::process;
 use std::str::FromStr;
 
 use nodelib::error::{Error, ResourceError};
@@ -18,38 +13,13 @@ use nodelib::resources::node::*;
 use nodelib::resources::pod::*;
 use nodelib::resources::Resource;
 
-/// Exit code to use if the server fails to start
-pub static EXIT_FAILURE: i32 = 1;
-
-pub async fn serve(etcd_config: etcd::Config, address: Ipv4Addr) {
-    if let Err(error) = run(etcd_config, address).await {
-        tracing::error!(?error, "could not serve application, terminating...");
-        process::exit(EXIT_FAILURE);
-    }
-}
-
-async fn run(etcd_config: etcd::Config, address: Ipv4Addr) -> std::io::Result<()> {
-    let app = Router::new()
-        .route("/resources", post(create_resource))
-        .route("/resources/:type", get(list_resources))
-        .route("/resources/:type/:name", get(get_resource))
-        .route("/resources/:type/:name", patch(patch_resource))
-        .route("/resources/:type/:name", delete(delete_resource))
-        .with_state(etcd_config);
-
-    let listener = tokio::net::TcpListener::bind((address, 80)).await?;
-    axum::serve(listener, app).await?;
-
-    Ok(())
-}
-
-///////////////////////////////////////////////////////////////////////////////
+use crate::web::errors::ProblemDetail;
 
 // TODO: return a `ProblemDetail` if the request fails to decode - see
 // https://github.com/tokio-rs/axum/blob/main/examples/customize-extractor-error/src/custom_extractor.rs
 
 /// Create a resource.
-async fn create_resource(
+pub async fn create(
     State(etcd_config): State<etcd::Config>,
     Json(payload): Json<Resource>,
 ) -> impl IntoResponse {
@@ -78,7 +48,7 @@ async fn create_resource(
 }
 
 /// List all resources of the given type.
-async fn list_resources(
+pub async fn list(
     State(etcd_config): State<etcd::Config>,
     Path(rtype): Path<String>,
 ) -> impl IntoResponse {
@@ -92,7 +62,7 @@ async fn list_resources(
 }
 
 /// Get a specific resource.
-async fn get_resource(
+pub async fn get(
     State(etcd_config): State<etcd::Config>,
     Path((rtype, rname)): Path<(String, String)>,
 ) -> impl IntoResponse {
@@ -107,7 +77,7 @@ async fn get_resource(
 }
 
 /// Modify a resource in place - the `type` and `name` cannot be modified.
-async fn patch_resource(
+pub async fn patch(
     State(etcd_config): State<etcd::Config>,
     Path((rtype, rname)): Path<(String, String)>,
     Json(payload): Json<Value>,
@@ -143,7 +113,7 @@ async fn patch_resource(
 }
 
 /// Delete a specific resource.
-async fn delete_resource(
+pub async fn delete(
     State(etcd_config): State<etcd::Config>,
     Path((rtype, rname)): Path<(String, String)>,
 ) -> impl IntoResponse {
@@ -162,96 +132,6 @@ async fn delete_resource(
             Ok(false) => ProblemDetail::resource_not_found(&rtype, &rname).into_response(),
             Err(error) => ProblemDetail::from_error(error).into_response(),
         }
-    }
-}
-
-///////////////////////////////////////////////////////////////////////////////
-
-/// An RFC7807 "problem detail" object.
-#[derive(Serialize)]
-struct ProblemDetail {
-    #[serde(rename = "type")]
-    rtype: String,
-    title: String,
-    detail: String,
-    status: u16,
-}
-
-impl ProblemDetail {
-    fn from_error(error: Error) -> (StatusCode, Json<Self>) {
-        match error {
-            Error::Resource(r) => Self::resource_invalid(&format!("{r:?}")),
-            Error::EtcdResponse(_)
-            | Error::FromUtf8(_)
-            | Error::Streaming(_)
-            | Error::TonicStatus(_)
-            | Error::TonicTransport(_) => Self::internal_server_error(&format!("{error:?}")),
-        }
-    }
-
-    /// TODO: rtype URI
-    fn internal_server_error(detail: &str) -> (StatusCode, Json<Self>) {
-        let status = StatusCode::INTERNAL_SERVER_ERROR;
-        let pd = Self {
-            rtype: "internal-server-error".to_string(),
-            title: "Something went badly wrong, check the logs.".to_string(),
-            detail: detail.to_string(),
-            status: status.as_u16(),
-        };
-
-        (status, Json(pd))
-    }
-
-    /// TODO: rtype URI
-    fn resource_already_exists(rtype: &str, rname: &str) -> (StatusCode, Json<Self>) {
-        let status = StatusCode::CONFLICT;
-        let pd = Self {
-            rtype: "resource-already-exists".to_string(),
-            title: "The given resource already exists.".to_string(),
-            detail: format!("The resource '{rname}' of type '{rtype}' already exists.  If you want to replace it, delete it first."),
-            status: status.as_u16(),
-        };
-
-        (status, Json(pd))
-    }
-
-    /// TODO: rtype URI
-    fn resource_invalid(detail: &str) -> (StatusCode, Json<Self>) {
-        let status = StatusCode::UNPROCESSABLE_ENTITY;
-        let pd = Self {
-            rtype: "resource-invalid".to_string(),
-            title: "The given resource is invalid.".to_string(),
-            detail: detail.to_string(),
-            status: status.as_u16(),
-        };
-
-        (status, Json(pd))
-    }
-
-    /// TODO: rtype URI
-    fn resource_not_found(rtype: &str, rname: &str) -> (StatusCode, Json<Self>) {
-        let status = StatusCode::NOT_FOUND;
-        let pd = Self {
-            rtype: "resource-not-found".to_string(),
-            title: "The given resource does not exist.".to_string(),
-            detail: format!("The resource '{rname}' of type '{rtype}' does not exist."),
-            status: status.as_u16(),
-        };
-
-        (status, Json(pd))
-    }
-
-    /// TODO: rtype URI
-    fn modification_not_allowed(detail: &str) -> (StatusCode, Json<Self>) {
-        let status = StatusCode::UNPROCESSABLE_ENTITY;
-        let pd = Self {
-            rtype: "modification-not-allowed".to_string(),
-            title: "The given change is not permitted.".to_string(),
-            detail: detail.to_string(),
-            status: status.as_u16(),
-        };
-
-        (status, Json(pd))
     }
 }
